Name changes and moved motor kubejs entrypoint
