+++ src/main/java/aztech/modern_industrialization/compat/kubejs/KubeJSProxy.java
@@ -55,6 +55,9 @@ public class KubeJSProxy {
     public void fireRegisterUpgradesEvent() {
     }
 
+    public void fireRegisterAndModifyMotorsEvents() {
+    }
+
     public void fireAddMultiblockSlotsEvent(String category, SlotPositions.Builder itemInputs, SlotPositions.Builder itemOutputs,
             SlotPositions.Builder fluidInputs, SlotPositions.Builder fluidOutputs) {
     }
