+++ src/main/java/aztech/modern_industrialization/compat/kubejs/LoadedKubeJSProxy.java
@@ -27,8 +27,9 @@ import aztech.modern_industrialization.compat.kubejs.machine.*;
 import aztech.modern_industrialization.compat.kubejs.material.AddMaterialsEventJS;
 import aztech.modern_industrialization.compat.kubejs.material.MIMaterialKubeJSEvents;
 import aztech.modern_industrialization.compat.kubejs.material.ModifyMaterialEventJS;
-import aztech.modern_industrialization.compat.kubejs.pipe.MIPipeKubeJSEvents;
-import aztech.modern_industrialization.compat.kubejs.pipe.ModifyMotorsEventJS;
+import aztech.modern_industrialization.compat.kubejs.motor.MIPipeKubeJSEvents;
+import aztech.modern_industrialization.compat.kubejs.motor.ModifyMotorsEventJS;
+import aztech.modern_industrialization.compat.kubejs.motor.RegisterMotorsEventJS;
 import aztech.modern_industrialization.compat.kubejs.recipe.CustomConditionEventJS;
 import aztech.modern_industrialization.compat.kubejs.recipe.MIRecipeKubeJSEvents;
 import aztech.modern_industrialization.compat.kubejs.registration.MIRegistrationKubeJSEvents;
@@ -80,6 +81,11 @@ public class LoadedKubeJSProxy extends KubeJSProxy {
     @Override
     public void fireRegisterUpgradesEvent() {
         MIMachineKubeJSEvents.REGISTER_UPGRADES.post(new RegisterUpgradesEventJS());
+    }
+
+    @Override
+    public void fireRegisterAndModifyMotorsEvents() {
+        MIPipeKubeJSEvents.REGISTER.post(new RegisterMotorsEventJS());
         MIPipeKubeJSEvents.MODIFY_MOTORS.post(new ModifyMotorsEventJS());
     }
